refactor(DelimiterInputData): 구분자와 입력 정보를 담는 dto 변수명 수정
